# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gpt-terminal-plus API",
    "version": "0.1.0",
    "description": "OpenAPI generated from JSDoc annotations.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Public base URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_TOKEN"
      }
    },
    "schemas": {
      "ExecutionResult": {
        "type": "object",
        "properties": {
          "stdout": {
            "type": "string"
          },
          "stderr": {
            "type": "string"
          },
          "exitCode": {
            "type": "integer"
          },
          "error": {
            "type": "boolean"
          }
        },
        "required": [
          "stdout",
          "stderr",
          "exitCode",
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/command/execute": {
      "post": {
        "operationId": "executeCommand",
        "summary": "Execute using first available mode",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "command": {
                    "type": "string"
                  }
                },
                "required": [
                  "command"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/ExecutionResult"
                    },
                    "aiAnalysis": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "result"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
    "/command/execute-llm": {
      "post": {
        "operationId": "executeLlm",
        "summary": "Run an LLM plan or direct instruction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "instructions": {
                    "type": "string"
                  },
                  "dryRun": {
                    "type": "boolean"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "instructions"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "LLM execution complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "object"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
  },
...
}